+++ Structural/Structural.Decorator/CoffeeExample/ConcreteComponents/TurkCoffee.cs
@@ -7,7 +7,7 @@
             Description = "House Blend Coffee";
         }
 
-        public override double Cost() => 0.89;
+        public override decimal Cost() => 0.89M;
 
         public override string GetDescription() => this.Description;
     }
